pyproject.toml
[tool.poetry]
name = "trader"
version = "0.1.0"
description = ""
authors = ["Anderson Ribeiro Lopes <zeander21@outlook.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.1.1"
sqlalchemy = "^2.0.3"
python-binance = "^1.0.16"
python-dotenv = "^0.21.1"
aiologger = "^0.7.0"
asyncpg = "^0.27.0"
pydantic = "^1.10.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
